Update to vscode

+++ src/PCF8575Debounce.cpp
@@ -18,7 +18,7 @@ void PCF8575Debounce::begin(uint8_t address, uint8_t intPin) {
 }
 
 uint16_t PCF8575Debounce::readPCF() {
-  Wire.requestFrom(address, 2);
+  Wire.requestFrom(address, static_cast<size_t>(2));
   uint16_t data = Wire.read();
   data |= (Wire.read() << 8);
   return data;
